+++ packages/nimble-components/src/wafer-map/modules/experimental/worker-renderer.ts
@@ -40,6 +40,7 @@ export class WorkerRenderer {
         if (this.wafermap.diesTable === undefined) {
             return;
         }
+
         await this.wafermap.worker.setTransform(this.wafermap.transform);
         const topLeftCanvasCorner = this.wafermap.transform.invert([0, 0]);
         const bottomRightCanvasCorner = this.wafermap.transform.invert([
@@ -99,8 +100,7 @@ export class WorkerRenderer {
     }
 
     private async setupWorker(): Promise<void> {
-        if (!this.wafermap.isExperimentalUpdate()
-        ) {
+        if (!this.wafermap.isExperimentalUpdate()) {
             return;
         }
         if (!this.isWorkerAlive) {